+++ frontend/src/app/users/users.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, ViewChild, TemplateRef } from '@angular/core';
 import { User } from './user';
 import { UserService } from './user.service';
 import { Router } from '@angular/router';
+import { Group } from './group';
 
 @Component({
   selector: 'users',
@@ -16,23 +17,36 @@ export class UsersComponent implements OnInit {
 
   editedUser: User;
   users: Array<User>;
-  isNewRecord: boolean;
-  statusMessage: string;
+  groups: Array<Group>;
 
   constructor(
     private serv: UserService,
     private router: Router
   ) {
     this.users = new Array<User>();
+    this.groups = new Array<Group>();
+    this.editedUser = null;
   }
 
   ngOnInit(): void {
-    this.loadUsers();
+    this.loadGroups();
+  }
+
+  private loadGroups() {
+    this.serv.getGroups().subscribe((data: Group[]) => {
+      this.groups = data;
+      this.loadUsers();
+    },
+    (error: any) => {
+      if (error.status == 401) {
+        this.router.navigate(['auth']);
+      }
+    })
   }
 
   private loadUsers() {
-    this.serv.getUsers().subscribe((data: User[]) => {
-      this.users = data
+      this.serv.getUsers().subscribe((data: User[]) => {
+      this.users = data;
     },
     (error: any) => {
       if (error.status == 401) {
@@ -42,45 +56,44 @@ export class UsersComponent implements OnInit {
   }
 
   editUser(user: User) {
-    this.editedUser = new User(user.id, user.username, user.roleId)
+    this.editedUser = new User(user.id, user.username, user.email, user.roleId, user.group_id);
   }
 
   loadTemplate(user: User) {
     if (this.editedUser && this.editedUser.id == user.id) {
-      return this.editTemplate
+      return this.editTemplate;
     } else {
-      return this.readOnlyTemplate
+      return this.readOnlyTemplate;
     }
   }
 
   saveUser() {
-    if (this.isNewRecord) {
-      this.serv.createUser(this.editedUser).subscribe(data => {
-        ;(this.statusMessage = 'Данные успешно добавлены'), this.loadUsers()
-      })
-      this.isNewRecord = false
-      this.editedUser = null
-    } else {
       this.serv
         .updateUser(this.editedUser.id, this.editedUser)
         .subscribe(data => {
-          ;(this.statusMessage = 'Данные успешно обновлены'), this.loadUsers()
+          setTimeout(() => this.loadUsers(), 1000);
+        },
+        (error: any) => {
+          if (error.status == 403) {
+            alert('Вы не можете редактировать текущего пользователя!');
+          }
         })
-      this.editedUser = null
-    }
+      this.editedUser = null;
   }
 
   cancel() {
-    if (this.isNewRecord) {
-      this.users.pop()
-      this.isNewRecord = false
-    }
-    this.editedUser = null
+    this.editedUser = null;
   }
 
   deleteUser(user: User) {
+    if (!confirm("Удалить пользователя?")) return;
     this.serv.deleteUser(user.id).subscribe(data => {
-      ;(this.statusMessage = 'Данные успешно удалены'), this.loadUsers()
+      setTimeout(() => this.loadUsers(), 1000);
+    },
+    (error: any) => {
+      if (error.status == 403) {
+        alert('Вы не можете удалить текущего пользователя!');
+      }
     })
   }
 }
